Clean up statistics tests
